refactor: 게시판 게시글 목록 페이지 코드 리팩토링 진행
