+++ src/widget/board/BoardHeader.tsx
@@ -1,30 +1,26 @@
 "use client";
 
-import { IconButton, PreviousButton, usePostListStore } from "@/shared";
+import { BoardService, IconButton, Loading, PreviousButton } from "@/shared";
 import { useParams, useRouter } from "next/navigation";
 
-export const PostListHeader = () => {
+export const BoardHeader = () => {
+  const params = useParams();
+  const { boardId } = params;
   const router = useRouter();
-  const param = useParams();
-  const { boardId } = param;
-  const initState = () => {
-    setPage(0);
-    setPosts([]);
-  };
-  const { boardName, setPage, setPosts } = usePostListStore();
+  const { useGetBoardName } = BoardService();
+  const { isSuccess, data } = useGetBoardName(boardId as string);
 
   return (
     <div className="flex h-24 w-full items-end px-5 sm:px-10">
       <PreviousButton />
       <div className="z-10 flex w-full items-center justify-between">
         <div className="truncate pr-4 text-xl font-bold lg:text-3xl">
-          {boardName}
+          {isSuccess ? data : ""}
         </div>
         <div className="flex items-center gap-2 sm:gap-4">
           <IconButton
             iconName={"add"}
             callback={() => {
-              initState();
               router.push(`/board/${boardId}/create`);
             }}
           />
@@ -36,7 +32,6 @@ export const PostListHeader = () => {
           <IconButton
             iconName={"search"}
             callback={() => {
-              initState();
               router.push(`/board/${boardId}/search`);
             }}
           />
